refactor: change dir for scroll-to-top.tsx
